Implement 'K_Cmd_CREATE_NPC_SYNC' + change return type of 'ScrCmdFunc' to u32

+++ include/Script/scrCommonCommand.h
@@ -3,7 +3,7 @@
 
 #include "Utils.h"
 
-typedef u8 (*ScrCmdFunc)();
+typedef u32 (*ScrCmdFunc)();
 
 // 8 bytes
 typedef struct
