+++ src/Script/scrCommonCommand.c
@@ -6,13 +6,13 @@
 #include "datCalendar.h"
 #include "temporary.h"
 
-u8 scrCommand_RAND();
-u8 scrCommand_SYNC();
-u8 scrCommand_WAIT();
-u8 scrCommand_PUT();
-u8 scrCommand_PUTSTR();
-u8 scrCommand_SWITCH();
-u8 scrCommand_SQRT();
+u32 scrCommand_RAND();
+u32 scrCommand_SYNC();
+u32 scrCommand_WAIT();
+u32 scrCommand_PUT();
+u32 scrCommand_PUTSTR();
+u32 scrCommand_SWITCH();
+u32 scrCommand_SQRT();
 
 // 007b92b0
 ScrCommandTable gScrCmdTable =
@@ -27,7 +27,7 @@ ScrCommandTable gScrCmdTable =
 };
 
 // FUN_0035afb0
-u8 scrCommand_RAND()
+u32 scrCommand_RAND()
 {
     scrSetIntReturnVal(RpRandom() % (scrGetIntPara(0) + 1));
 
@@ -35,13 +35,13 @@ u8 scrCommand_RAND()
 }
 
 // FUN_0035b000
-u8 scrCommand_SYNC()
+u32 scrCommand_SYNC()
 {
     return scrGetCmdTimer() != 0;
 }
 
 // FUN_0035b030
-u8 scrCommand_WAIT()
+u32 scrCommand_WAIT()
 {
     u32 cmdTimer;
 
@@ -56,7 +56,7 @@ u8 scrCommand_WAIT()
 }
 
 // FUN_0035b090
-u8 scrCommand_PUT()
+u32 scrCommand_PUT()
 {
     s32 param;
 
@@ -67,7 +67,7 @@ u8 scrCommand_PUT()
 }
 
 // FUN_0035b0d0
-u8 scrCommand_PUTSTR()
+u32 scrCommand_PUTSTR()
 {
     char* param;
 
@@ -78,7 +78,7 @@ u8 scrCommand_PUTSTR()
 }
 
 // FUN_0035b110
-u8 scrCommand_SWITCH()
+u32 scrCommand_SWITCH()
 {
     s32 caseIdx;
     s32 lblIdx;
@@ -103,7 +103,7 @@ u8 scrCommand_SWITCH()
 }
 
 // FUN_0035b520
-u8 scrCommand_SQRT()
+u32 scrCommand_SQRT()
 {
     f32 param;
